requirements.txt
# Core UI
streamlit>=1.32.0

# Vector database
chromadb>=0.5.0

# LangChain ecosystem
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-openai>=0.1.0

# Local embeddings (no API key required)
sentence-transformers>=3.0.0

# PDF parsing
pypdf>=4.0.0
PyMuPDF>=1.24.0

# Data validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
